+++ Winey-API/src/main/java/com/example/wineyapi/tastingNote/service/TastingNoteService.java
@@ -13,4 +13,6 @@ public interface TastingNoteService {
     TastingNoteResponse.CreateTastingNoteDTO createTastingNote(User user, TastingNoteRequest.CreateTastingNoteDTO request);
 
     PageResponse<List<TastingNoteResponse.TastingNoteListDTO>> getTastingNoteList(User user, Integer page, Integer size, Integer order, List<Country> countries, List<WineType> wineTypes, Integer isBuyAgain);
+
+    TastingNoteResponse.TasteAnalysisDTO tasteAnalysis(User user);
 }
