+++ Winey-API/src/main/java/com/example/wineyapi/wine/dto/WineResponse.java
@@ -58,112 +58,6 @@ public class WineResponse {
     @Setter
     @AllArgsConstructor
     @Builder
-    public static class TasteAnalysisDTO {
-        @Schema(name = "추천 국가")
-        private String recommendCountry;
-
-        @Schema(name = "추천 품종")
-        private String recommendVarietal;
-
-        @Schema(name = "추천 와인타입")
-        private String recommendWineType;
-
-        @Schema(name = "총 마신 와인 병수")
-        private int totalWineCnt;
-
-        @Schema(name = "재구매 의사가 있는 와인")
-        private int buyAgainCnt;
-
-        @Schema(name = "레드 와인 갯수")
-        private int redCnt;
-
-        @Schema(name = "화이트 와인 갯수")
-        private int whiteCnt;
-
-        @Schema(name = "스파클링 와인 갯수")
-        private int sparklingCnt;
-
-        @Schema(name = "로제와인 갯수")
-        private int roseCnt;
-
-        @Schema(name = "주정강화 갯수")
-        private int fortifiedCnt;
-
-        @Schema(name = "기타 와인 갯수")
-        private int otherCnt;
-
-        @Schema(description = "key = country, value = percent 식인 List 입니다.")
-        private List<Top3Country> top3Country;
-
-        @Schema(description = "key = varietal, value = percent 식인 List 입니다.")
-        private List<Top3Varietal> top3Varietal;
-
-        @Schema(description = "key = smell, value = percent 식인 List 입니다. ")
-        private List<Top7Smell> top7Smell;
-
-        private Taste taste;
-
-        @Schema(name = "평균 와인 가격")
-        private int avgPrice;
-
-    }
-
-    @NoArgsConstructor
-    @Getter
-    @Setter
-    @AllArgsConstructor
-    @Builder
-    public static class Taste{
-        private double sweetness;
-
-        private double acidity;
-
-        private double alcohol;
-
-        private double body;
-
-        private double tannin;
-
-        private double finish;
-    }
-
-    @NoArgsConstructor
-    @Getter
-    @Setter
-    @AllArgsConstructor
-    @Builder
-    public static class Top3Country{
-        private String country;
-
-        private int percent;
-    }
-
-    @NoArgsConstructor
-    @Getter
-    @Setter
-    @AllArgsConstructor
-    @Builder
-    public static class Top3Varietal{
-        private String varietal;
-
-        private int percent;
-    }
-    @NoArgsConstructor
-    @Getter
-    @Setter
-    @AllArgsConstructor
-    @Builder
-    public static class Top7Smell{
-        private String smell;
-
-        private int percent;
-    }
-
-    @NoArgsConstructor
-    @Getter
-    @Setter
-    @AllArgsConstructor
-    @Builder
     public static class SearchWineDto {
         @Schema(name = "와인 id")
         private Long wineId;
