+++ Winey-API/src/main/java/com/example/wineyapi/wine/service/WineService.java
@@ -10,7 +10,5 @@ import java.util.List;
 public interface WineService {
     List<WineResponse.RecommendWineDTO> recommendWine(User user);
 
-    WineResponse.TasteAnalysisDTO tasteAnalysis(User user);
-
     PageResponse<List<WineResponse.SearchWineDto>> searchWineList(Integer page, Integer size, String content);
 }
